feat(builder)t: make the client bundler takes multiple entries at once
