+++ crates/metassr-build/Cargo.toml
@@ -2,9 +2,10 @@
 name = "metassr-build"
 version = "0.1.0"
 edition = "2021"
-description = "Web builder & bundler built on Rspack and built for MetaSSR."
+description = "Web builder & client bundler (built on Rspack), built for MetaSSR."
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 metacall = "0.4.0"
 anyhow = "1.0.82"
+serde_json = "1.0.120"
