+++ crates/metassr-build/src/client/mod.rs
@@ -1,0 +1,1 @@
+pub mod bundler;
